DocumentsSection - move SectionContainer and rename DocumentsAllSection to DocumentsAll for consistency
